# tp4-ecommerce/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopecart E-commerce API Documentation",
        "description": "Documentation des points d'accès (endpoints) de l'API Shopecart.",
        "contact": {
            "email": "contact@shopecart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de l'API locale"
        }
    ],
    "paths": {
...
        "/api/cartItems/cart/{cartId}": {
            "get": {
                "tags": [
                    "CartItem"
                ],
                "summary": "Get all cart items for a specific cart",
                "operationId": "ca52ae915104f3d82d08b03c601cee47",
                "parameters": [
                    {
                        "name": "cartId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cart items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "cartId": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "productVariantId": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/api/cartItems/{cartItemId}": {
            "get": {
                "tags": [
                    "CartItem"
                ],
                "summary": "Get a single cart item by ID",
                "operationId": "6656432577c7e26ffb9b88cf18ddcebc",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "description": "ID of the cart item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart item retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cartId": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "productVariantId": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-13T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-13T12:05:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cart item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "CartItem"
                ],
                "summary": "Update an existing cart item",
                "operationId": "959220efcd68eb08d14cc121a69cddb1",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "description": "ID of the cart item to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "cartId",
                                    "productVariantId"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "cartId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "productVariantId": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart item updated successfully"
                                        },
                                        "cartItem": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cartId": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "productVariantId": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-13T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-13T12:05:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cart item or product variant not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CartItem"
                ],
                "summary": "Delete a cart item",
                "operationId": "9282fe541984a1b64b350a44d643e678",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "description": "ID of the cart item to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart item deleted successfully"
                    },
                    "404": {
                        "description": "Cart item not found"
                    }
                }
            }
        },
        "/api/cartItems": {
            "post": {
                "tags": [
                    "CartItem"
                ],
                "summary": "Add a new cart item",
                "operationId": "dc9045ea47ded10c062f6c971ae2b4aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "cartId",
                                    "productVariantId"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "cartId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "productVariantId": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cartId": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "productVariantId": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-13T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-13T12:05:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product variant not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le jeton Bearer obtenu après la connexion.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}